fix(BA-5707): separate prometheus setup module to prevent import-time ValueClass misfire
prometheus_client evaluates ValueClass once at import time based on
PROMETHEUS_MULTIPROC_DIR env var. The old multiprocess.py imported
prometheus_client at module top level, so importing setup_prometheus_multiprocess_dir
also triggered ValueClass evaluation before the env var was set — silently
fixing it as MutexValue (in-memory only) instead of MmapedValue (mmap file).

On macOS (spawn) workers re-import fresh so the bug was masked; on Linux
(fork) workers inherit the parent's module cache with the wrong ValueClass,
causing all Prometheus metrics to silently vanish.

Split multiprocess_setup.py (no prometheus_client import) from
multiprocess.py (prometheus_client-dependent helpers) so the env var is
always set before the library loads.
