+++ src/ai/backend/account_manager/cli/start_server.py
@@ -18,7 +18,7 @@ def main(ctx: click.Context) -> None:
 
     This is a thin wrapper that defers the heavy import of server module.
     """
-    from ai.backend.common.metrics.multiprocess import setup_prometheus_multiprocess_dir
+    from ai.backend.common.metrics.multiprocess_setup import setup_prometheus_multiprocess_dir
 
     setup_prometheus_multiprocess_dir("account-manager")
