+++ tests/unit/common/test_prometheus_multiprocess.py
@@ -8,10 +8,10 @@ from unittest.mock import patch
 
 import pytest
 
-import ai.backend.common.metrics.multiprocess as mp_mod
-from ai.backend.common.metrics.multiprocess import (
+import ai.backend.common.metrics.multiprocess_setup as mp_setup_mod
+from ai.backend.common.metrics.multiprocess import generate_latest_multiprocess
+from ai.backend.common.metrics.multiprocess_setup import (
     cleanup_prometheus_multiprocess_dir,
-    generate_latest_multiprocess,
     setup_prometheus_multiprocess_dir,
 )
 
@@ -19,10 +19,10 @@ from ai.backend.common.metrics.multiprocess import (
 @pytest.fixture(autouse=True)
 def _reset_multiprocess_state() -> Iterator[None]:
     """Reset the module-level global state before each test."""
-    original_dir = mp_mod._multiprocess_dir
+    original_dir = mp_setup_mod._multiprocess_dir
     original_env = os.environ.get("PROMETHEUS_MULTIPROC_DIR")
     yield
-    mp_mod._multiprocess_dir = original_dir
+    mp_setup_mod._multiprocess_dir = original_dir
     if original_env is not None:
         os.environ["PROMETHEUS_MULTIPROC_DIR"] = original_env
     elif "PROMETHEUS_MULTIPROC_DIR" in os.environ:
@@ -31,7 +31,7 @@ def _reset_multiprocess_state() -> Iterator[None]:
 
 class TestSetupPrometheusMultiprocDir:
     def test_creates_directory_with_default_base(self, tmp_path: Path) -> None:
-        with patch.object(mp_mod, "_DEFAULT_BASE_DIR", tmp_path):
+        with patch.object(mp_setup_mod, "_DEFAULT_BASE_DIR", tmp_path):
             result = setup_prometheus_multiprocess_dir("manager")
 
         assert result == tmp_path / "manager"
@@ -45,7 +45,7 @@ class TestSetupPrometheusMultiprocDir:
         (prom_dir / "counter_456.db").touch()
         (prom_dir / "keep_this.txt").touch()
 
-        with patch.object(mp_mod, "_DEFAULT_BASE_DIR", tmp_path):
+        with patch.object(mp_setup_mod, "_DEFAULT_BASE_DIR", tmp_path):
             setup_prometheus_multiprocess_dir("manager")
 
         assert not (prom_dir / "gauge_liveall_123.db").exists()
@@ -53,7 +53,7 @@ class TestSetupPrometheusMultiprocDir:
         assert (prom_dir / "keep_this.txt").exists()
 
     def test_idempotent_returns_same_path(self, tmp_path: Path) -> None:
-        with patch.object(mp_mod, "_DEFAULT_BASE_DIR", tmp_path):
+        with patch.object(mp_setup_mod, "_DEFAULT_BASE_DIR", tmp_path):
             first = setup_prometheus_multiprocess_dir("manager")
             second = setup_prometheus_multiprocess_dir("manager")
 
@@ -61,7 +61,7 @@ class TestSetupPrometheusMultiprocDir:
 
     def test_idempotent_ignores_different_component(self, tmp_path: Path) -> None:
         """Once setup, calling again with different component still returns the first path."""
-        with patch.object(mp_mod, "_DEFAULT_BASE_DIR", tmp_path):
+        with patch.object(mp_setup_mod, "_DEFAULT_BASE_DIR", tmp_path):
             first = setup_prometheus_multiprocess_dir("manager")
             second = setup_prometheus_multiprocess_dir("agent")
 
@@ -97,7 +97,7 @@ class TestGenerateLatestMultiprocess:
         prom_dir = tmp_path / "test-component"
         prom_dir.mkdir(parents=True)
         os.environ["PROMETHEUS_MULTIPROC_DIR"] = str(prom_dir)
-        mp_mod._multiprocess_dir = prom_dir
+        mp_setup_mod._multiprocess_dir = prom_dir
 
         result = generate_latest_multiprocess()
         assert isinstance(result, bytes)
@@ -106,7 +106,7 @@ class TestGenerateLatestMultiprocess:
         prom_dir = tmp_path / "test-component"
         prom_dir.mkdir(parents=True)
         os.environ["PROMETHEUS_MULTIPROC_DIR"] = str(prom_dir)
-        mp_mod._multiprocess_dir = prom_dir
+        mp_setup_mod._multiprocess_dir = prom_dir
 
         # Simulate systemd-tmpfiles-clean deleting the directory
         prom_dir.rmdir()
@@ -119,7 +119,7 @@ class TestGenerateLatestMultiprocess:
 
     def test_returns_empty_bytes_on_unrecoverable_failure(self) -> None:
         # Set an invalid path that can't be recreated
-        mp_mod._multiprocess_dir = Path("/nonexistent/impossible/path")
+        mp_setup_mod._multiprocess_dir = Path("/nonexistent/impossible/path")
         os.environ["PROMETHEUS_MULTIPROC_DIR"] = "/nonexistent/impossible/path"
 
         result = generate_latest_multiprocess()
@@ -132,16 +132,16 @@ class TestCleanupPrometheusMultiprocDir:
         prom_dir.mkdir(parents=True)
         (prom_dir / "test.db").touch()
         os.environ["PROMETHEUS_MULTIPROC_DIR"] = str(prom_dir)
-        mp_mod._multiprocess_dir = prom_dir
+        mp_setup_mod._multiprocess_dir = prom_dir
 
         cleanup_prometheus_multiprocess_dir()
 
         assert not prom_dir.exists()
         assert "PROMETHEUS_MULTIPROC_DIR" not in os.environ
-        assert mp_mod._multiprocess_dir is None
+        assert mp_setup_mod._multiprocess_dir is None
 
     def test_noop_when_not_initialized(self) -> None:
-        mp_mod._multiprocess_dir = None
+        mp_setup_mod._multiprocess_dir = None
         if "PROMETHEUS_MULTIPROC_DIR" in os.environ:
             del os.environ["PROMETHEUS_MULTIPROC_DIR"]
 
@@ -151,13 +151,13 @@ class TestCleanupPrometheusMultiprocDir:
 
 class TestDefaultBaseDirUid:
     def test_default_base_dir_is_hardcoded(self) -> None:
-        assert Path("./run/prometheus") == mp_mod._DEFAULT_BASE_DIR
+        assert Path("./run/prometheus") == mp_setup_mod._DEFAULT_BASE_DIR
 
     def test_setup_raises_on_permission_error(
         self, tmp_path: Path, caplog: pytest.LogCaptureFixture
     ) -> None:
         with (
-            patch.object(mp_mod, "_DEFAULT_BASE_DIR", tmp_path),
+            patch.object(mp_setup_mod, "_DEFAULT_BASE_DIR", tmp_path),
             patch.object(Path, "mkdir", side_effect=PermissionError("mock denied")),
             caplog.at_level(logging.ERROR),
         ):
